cli/cmd/daemon: fix duration flag
Since duration by default is in nanoseconds, the CLI flag help text said
'300ns' for the default but we treated it as seconds. Use an actual
`time.Duration` object set to 300s as the default to align the help text
and actual configuration.

Users can configure any value that can be parsed by
`time.ParseDuration`.