pkg/daemon: switch to default RSA key size
There's no reason to use a stronger key than the default, as the chain contains
other RSA2048 keys. Together with the shorter preferred chain, this should
reduce the size of the TLS handshake a bit and ensure we fit in 4096 byte
buffer that iPXE has for the TLS handshake record.